Accounts-Migrations

+++ app/routes/tenant.py
@@ -1,5 +1,5 @@
-import uuid
 from typing import TYPE_CHECKING
+from uuid import UUID
 
 from better_profanity import profanity
 from fastapi import APIRouter, BackgroundTasks, Depends, Path, Query
@@ -8,10 +8,10 @@ from pydantic import EmailStr
 from temporalio.client import WorkflowHandle
 
 from app.core.db import DBManager, get_db_manager
-from app.core.ijson import ijson_dumps
+from app.core.ijson import ijson_dumps, ijson_loads
 from app.core.oauth2 import get_oauth_scheme
 from app.exceptions import errors
-from app.models.input_param_patterns import TENANT_NAME_PATTERN
+from app.models.input_param_patterns import TENANT_NAME_PATTERN, TOKEN_PATTERN
 from app.models.product import ProductEnum
 from app.models.tenant import (
     SuggestTenantNamesResponseModel,
@@ -19,8 +19,10 @@ from app.models.tenant import (
     TenantResponseModel,
     TenantStatusEnum,
 )
-from app.route_utils.product import validate_email_domain
+from app.route_utils.product import validate_email_domain, validate_provisioning_details
+from app.route_utils.session_util import get_treated_email
 from app.route_utils.tenant_suggestions import get_existing_tenant_names, get_valid_suggestions
+from app.route_utils.user_session import UserSession
 
 if TYPE_CHECKING:
     from app.cli.base_workflow import ProductWorkflow
@@ -49,7 +51,7 @@ async def create_tenant(tenant_details: TenantCreateRequestModel) -> dict:
     response_model=list[TenantResponseModel] | None,
 )
 async def list_tenants(
-    product: ProductEnum = Path(...), tenant_id: uuid.UUID | None = None, _: dict = Depends(get_oauth_scheme())
+    product: ProductEnum = Path(...), tenant_id: UUID | None = None, _: dict = Depends(get_oauth_scheme())
 ) -> list[TenantResponseModel]:
     """
     @param product:
@@ -75,12 +77,12 @@ async def update_provisioning_workflow(product: ProductEnum, tenant_details: dic
     product_workflow: ProductWorkflow = ProductEnum.get_class(product)()
 
     workflow_handle: WorkflowHandle = await product_workflow.get_workflow_handle(schema=tenant_details)
+    if workflow_handle:
+        response = await workflow_handle.describe()
 
-    response = await workflow_handle.describe()
-
-    if response.status.name == "RUNNING":
-        # terminate the workflow
-        await workflow_handle.terminate()
+        if response.status.name == "RUNNING":
+            # terminate the workflow
+            await workflow_handle.terminate()
 
     # start the workflow
     await product_workflow.onboard(tenant_details)
@@ -89,13 +91,13 @@ async def update_provisioning_workflow(product: ProductEnum, tenant_details: dic
 
 
 @tenant_router.put(
-    "/updateTenantDetails/{product}",
+    "/{product}/updateTenantDetails/{tenant_id}",
     operation_id="updateTenantDetails",
 )
 async def update_tenant(
-    tenant_id: str,
     tenant_details: dict,
     product: ProductEnum = Path(..., alias="product"),
+    tenant_id: UUID = Path(...),
     background_tasks: BackgroundTasks = BackgroundTasks(),
     _: dict = Depends(get_oauth_scheme()),
 ) -> dict:
@@ -108,15 +110,18 @@ async def update_tenant(
     @return:
     """
     db: DBManager = await get_db_manager()
-    tenant = await db.fetch_one("get_tenant.sql", tenant_id=tenant_id)
+    tenant = await db.fetch_one("get_tenant.sql", tenant_id=str(tenant_id))
     if not tenant:
         raise errors.TENANT_NOT_FOUND.exc()
 
     if not TenantStatusEnum.can_update_tenant(TenantStatusEnum(tenant["status"])):
         raise errors.TENANT_DETAILS_CANNOT_BE_UPDATED.exc()
 
+    schema = ijson_loads(tenant["schema"]) if tenant["schema"] else None
+    await validate_provisioning_details(product=product, data=tenant_details, schema=schema)
+
     tenant_details_str = ijson_dumps(tenant_details)
-    response = await db.fetch_one("update_tenant_details.sql", schema_=tenant_details_str, tenant_id=tenant_id)
+    response = await db.fetch_one("update_tenant_details.sql", data=tenant_details_str, tenant_id=f"'{tenant_id!s}'")
 
     background_tasks.add_task(update_provisioning_workflow, product, tenant_details)
 
@@ -124,21 +129,23 @@ async def update_tenant(
 
 
 @tenant_router.get(
-    "/suggestTenantNames/{product}",
+    "/{product}/suggestTenantNames",
     operation_id="suggestTenantNames",
 )
 async def suggest_tenant_names(
     product: ProductEnum = Path(...),
     email: EmailStr = Query(...),
     organization: str = Query(...),
-    _: dict = Depends(get_oauth_scheme()),
+    token: str = Query(..., pattern=TOKEN_PATTERN),
 ) -> SuggestTenantNamesResponseModel:
     """
     @param organization:
     @param product:
     @return:
     """
+    email = get_treated_email(email)
     validate_email_domain(product=product, email=email)
+    await UserSession.validate_session(product=product, email=email, session_token=token)
     tenant_names = await get_valid_suggestions(product=product, email=email, organization=organization)
     return SuggestTenantNamesResponseModel(
         tenant_names=tenant_names,
@@ -154,6 +161,31 @@ async def verify_tenant_name(
     tenant_name: str = Path(pattern=TENANT_NAME_PATTERN),
     product: ProductEnum = Path(...),
     email: EmailStr = Query(...),
+    token: str = Query(..., pattern=TOKEN_PATTERN),
+) -> None:
+    """
+    @param product:
+    @param tenant_name:
+    @return:
+    """
+    email = get_treated_email(email)
+    validate_email_domain(product=product, email=email)
+    await UserSession.validate_session(product=product, email=email, session_token=token)
+    tenant_names = await get_existing_tenant_names(product=product, email=email, tenant_names=[tenant_name.lower()])
+    if tenant_names:
+        raise errors.ALREADY_ALLOCATED_TENANT_NAME.exc()
+    if profanity.contains_profanity(tenant_name):
+        raise errors.EXPLICIT_WORDS_NOT_ALLOWED.exc()
+
+
+@tenant_router.get(
+    "/{product}/keycloak/validateTenantName/{tenant_name}",
+    operation_id="validateTenantNameByKeycloak",
+)
+async def verify_tenant_name_by_keycloak(
+    tenant_name: str = Path(pattern=TENANT_NAME_PATTERN),
+    product: ProductEnum = Path(...),
+    email: EmailStr = Query(...),
     _: dict = Depends(get_oauth_scheme()),
 ) -> None:
     """
@@ -161,6 +193,7 @@ async def verify_tenant_name(
     @param tenant_name:
     @return:
     """
+    email = get_treated_email(email)
     validate_email_domain(product=product, email=email)
     tenant_names = await get_existing_tenant_names(product=product, email=email, tenant_names=[tenant_name.lower()])
     if tenant_names:
